+++ test/unit_tests/braket/aws/common_test_utils.py
@@ -17,12 +17,11 @@ from unittest.mock import Mock
 from braket.aws import AwsQuantumTaskBatch
 
 DWAVE_ARN = "arn:aws:braket:::device/qpu/d-wave/Advantage_system1"
-RIGETTI_ARN = "arn:aws:braket:::device/qpu/rigetti/Ankaa-2"
+RIGETTI_ARN = "arn:aws:braket:::device/qpu/rigetti/Cepheus-1-108Q"
 IONQ_ARN = "arn:aws:braket:us-east-1::device/qpu/ionq/Harmony"
 OQC_ARN = "arn:aws:braket:eu-west-2::device/qpu/oqc/Lucy"
 SV1_ARN = "arn:aws:braket:::device/quantum-simulator/amazon/sv1"
 DM1_ARN = "arn:aws:braket:::device/quantum-simulator/amazon/dm1"
-TN1_ARN = "arn:aws:braket:::device/quantum-simulator/amazon/tn1"
 XANADU_ARN = "arn:aws:braket:us-east-1::device/qpu/xanadu/Borealis"
 
 RIGETTI_REGION = "us-west-1"